Fixes the instanceOfOp and checkCastOp
Fixes the logic of these operators.
CheckCastOp now checks for the cast if the address is __not__ null.
InstanceOfOp checks if the address is an instance of another class if
the address is __not__ null.

In addition loads the class for the instanceof operator since the class
might not have been loaded yet.